+++ CliMenu/Program.cs
@@ -1,0 +1,6 @@
+using CliMenu.Components.DefaultComponents;
+
+Console.Clear();
+
+Bar bar = new("bar", 20);
+Console.WriteLine(bar.GetDisplay());